+++ app/api/movimentacao/[id]/route.ts
@@ -1,10 +1,17 @@
 import { prisma } from "@/app/lib/prisma";
 import { NextResponse } from "next/server";
+import { verifyToken } from "@/app/middleware/auth";
+import { authorize } from "@/app/middleware/role";
+import { Usuario } from "@/app/generated/prisma/client";
 
 export async function GET(
   req: Request,
   { params }: { params: Promise<{ id: string }> },
 ) {
+
+  const user = await verifyToken(req) as Usuario; 
+  authorize(user.role, ["OPERADOR", "ADMIN"]);
+
   const item = await prisma.movimentacao.findUnique({
     where: { id: Number((await params).id) },
   });
@@ -17,6 +24,10 @@ export async function PUT(
   req: Request,
   { params }: { params: Promise<{ id: string }> },
 ) {
+    
+  const user = await verifyToken(req) as Usuario; 
+  authorize(user.role, ["ADMIN"]);
+
   const body = await req.json();
 
   const atualizado = await prisma.movimentacao.update({
@@ -32,6 +43,9 @@ export async function DELETE(
   req: Request,
   { params }: { params: Promise<{ id: string }> },
 ) {
+  const user = await verifyToken(req) as Usuario; 
+  authorize(user.role, ["ADMIN"]);
+
   await prisma.movimentacao.delete({
     where: { id: Number((await params).id) },
   });
